rf: move `output_dandi_linkml_schema()` and supporting def
From `cli/tools.py` to `tools.py`
